requirements.txt
pynrrd>=1.0.0
matplotlib>=3.5.0
numpy>=1.21.0
